+++ examples/mnist.rs
@@ -6,7 +6,7 @@ const MAX_TRAIN_LENGHT: usize = 1000;
 const MAX_TEST_LENGHT: usize = 500;
 const EPOCHS: usize = 100;
 
-fn load_mnist() -> (Array2<f64>, Array2<f64>, Array2<f64>, Array2<f64>) {
+fn load_mnist() -> (Array2<f32>, Array2<f32>, Array2<f32>, Array2<f32>) {
     let Mnist {
         trn_img,
         trn_lbl,
@@ -23,19 +23,19 @@ fn load_mnist() -> (Array2<f64>, Array2<f64>, Array2<f64>, Array2<f64>) {
 
     let train_data = Array2::from_shape_vec((MAX_TRAIN_LENGHT, 28 * 28), trn_img)
         .expect("Error converting images to Array2 struct")
-        .map(|x| *x as f64 / 256.0);
+        .map(|x| *x as f32 / 256.0);
 
     let train_labels = Array2::from_shape_vec((MAX_TRAIN_LENGHT, 10), trn_lbl)
         .expect("Error converting training labels to Array2 struct")
-        .map(|x| *x as f64);
+        .map(|x| *x as f32);
 
     let test_data = Array2::from_shape_vec((MAX_TEST_LENGHT, 28 * 28), tst_img)
         .expect("Error converting images to Array2 struct")
-        .map(|x| *x as f64 / 256.);
+        .map(|x| *x as f32 / 256.);
 
     let test_labels = Array2::from_shape_vec((MAX_TEST_LENGHT, 10), tst_lbl)
         .expect("Error converting testing labels to Array2 struct")
-        .map(|x| *x as f64);
+        .map(|x| *x as f32);
 
     (train_data, train_labels, test_data, test_labels)
 }
